third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbobjectstore_createIndex6-event_order-expected.txt

This is a testharness.js-based test.
PASS IDBObjectStore.createIndex() - event order when unique constraint is triggered 
Harness: the test ran to completion.
